+++ scratch.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-console, no-unused-vars */
-import nlp from './src/two.js'
+import nlp from './src/three.js'
 // import nlp from './builds/compromise.cjs'
 // import text from '/Users/spencer/mountain/compromise/scripts/perf/flame/_sotu-text.js'
 
@@ -14,15 +14,16 @@ import nlp from './src/two.js'
 // console.log(doc.text())
 
 // let doc = nlp(`Pursuing a successful career, along with the usual social and financial advantages, will be easier this year `)
-// let doc = nlp(`all the days since december were gross`)
-let doc = nlp(`one two three`)
-doc.replace('two', 'foo bar')
-doc.debug()
+let doc = nlp(`all the days since december were gross`)
+
+// let doc = nlp(`one two three`)
+// doc.replace('two', 'foo bar')
+// doc.debug()
 
 // let doc = nlp(text)
 // doc.debug('chunks')
 // doc.chunks().debug()
-// doc.verbs().subjects().debug()
+doc.verbs().subjects().debug()
 // console.log(doc.nouns().json())
 //
 /*