+++ src/3-three/numbers/fractions/api.js
@@ -22,7 +22,7 @@ const plugin = function (View) {
     }
     json(n) {
       return getNth(this, n).map(p => {
-        let json = p.json()[0]
+        let json = p.toView().json()[0]
         let parsed = parse(p)
         json.fraction = parsed
         return json