+++ src/3-three/subjects/places/api.js
@@ -4,19 +4,11 @@ import find from './find.js'
 export const getNth = (doc, n) => (typeof n === 'number' ? doc.eq(n) : doc)
 
 const addMethod = function (View) {
-  /**
-   */
-  class Places extends View {
-    constructor(document, pointer, groups) {
-      super(document, pointer, groups)
-      this.viewType = 'Places'
-    }
-  }
 
   View.prototype.places = function (n) {
     let m = find(this)
     m = getNth(m, n)
-    return new Places(this.document, m.pointer)
+    return new View(this.document, m.pointer)
   }
 }
 export default addMethod